+++ HISTORY.rst
@@ -3,7 +3,17 @@
 History
 -------
 
-0.1.0 (2016-06-07)
+0.1.0 (2016-07-18)
 ++++++++++++++++++
 
-* First release on PyPI.
+* MoonInfo object that provides basic lunar information
+
+  * age
+  * altitude
+  * azimuth
+  * colongitude
+  * fractional phase
+  * libration latitude
+  * libration longitude
+  * next four lunar phases
+